// src/lib/database.types.ts
export interface Database {
  public: {
    Tables: {
      alunos: {
        Row: {
          cpf: string;
          created_at: string;
          email: string | null;
          id: string;
          id_responsavel: string | null;
          nome: string;
          plano: Database['public']['Enums']['planos'] | null;
          telefone: string | null;
          tem_responsavel: boolean;
        };
        Insert: {
          cpf: string;
          created_at?: string;
          email?: string | null;
          id?: string;
          id_responsavel?: string | null;
          nome: string;
          plano?: Database['public']['Enums']['planos'] | null;
          telefone?: string | null;
          tem_responsavel: boolean;
        };
        Update: {
          cpf?: string;
          created_at?: string;
          email?: string | null;
          id?: string;
          id_responsavel?: string | null;
          nome?: string;
          plano?: Database['public']['Enums']['planos'] | null;
          telefone?: string | null;
          tem_responsavel?: boolean;
        };
        Relationships: [
          {
            foreignKeyName: 'alunos_id_responsavel_fkey';
            columns: ['id_responsavel'];
            isOneToOne: false;
            referencedRelation: 'responsaveis';
            referencedColumns: ['id'];
          },
        ];
      };
// ...
      pagamentos: {
        Row: {
          aluno_id: string | null;
          created_at: string;
          data_pagamento: string | null;
          id: string;
          preco: number | null;
          tipo: Database['public']['Enums']['planos'];
          vigencia: string;
        };
        Insert: {
          aluno_id?: string | null;
          created_at?: string;
          data_pagamento?: string | null;
          id?: string;
          preco?: number | null;
          tipo: Database['public']['Enums']['planos'];
          vigencia: string;
        };
        Update: {
          aluno_id?: string | null;
          created_at?: string;
          data_pagamento?: string | null;
          id?: string;
          preco?: number | null;
          tipo?: Database['public']['Enums']['planos'];
          vigencia?: string;
        };
        Relationships: [
          {
            foreignKeyName: 'pagamentos_aluno_id_fkey';
            columns: ['aluno_id'];
            isOneToOne: false;
            referencedRelation: 'alunos';
            referencedColumns: ['id'];
          },
        ];
      };
// ...
      responsaveis: {
        Row: {
          cpf: string;
          created_at: string;
          email: string | null;
          id: string;
          nome: string;
          telefone: string | null;
        };
        Insert: {
          cpf: string;
          created_at?: string;
          email?: string | null;
          id?: string;
          nome: string;
          telefone?: string | null;
        };
        Update: {
          cpf?: string;
          created_at?: string;
          email?: string | null;
          id?: string;
          nome?: string;
          telefone?: string | null;
        };
        Relationships: [];
      };
// ...
    };
    Views: {
      [_ in never]: never;
    };
    Functions: {
      is_admin_user: {
        Args: {
          user_id: string;
        };
        Returns: boolean;
      };
      is_professor_user: {
        Args: {
          user_id: string;
        };
        Returns: boolean;
      };
    };
    Enums: {
      planos: 'mensal' | 'trimestral' | 'semestral' | 'anual';
      role: 'admin' | 'teacher';
    };
    CompositeTypes: {
      [_ in never]: never;
    };
  };
}